added translation layer for the yaml
tests needs to be completed later
